apps/dhg-baseline/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase-py==2.3.1
pydantic==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.26.0
postgrest==0.13.0
python-multipart==0.0.6
typing-extensions>=4.8.0
pydantic-settings==2.1.0 